feat: Add controllers

+++ blog-system-component/src/main/java/cc/lq/blog/system/service/impl/ArticleServiceImpl.java
@@ -1,6 +1,7 @@
 package cc.lq.blog.system.service.impl;
 
 import cc.lq.blog.system.entity.*;
+import cc.lq.blog.system.exception.ResourceNotFoundException;
 import cc.lq.blog.system.mapper.ArticleMapper;
 import cc.lq.blog.system.mapper.ArticleTagMapper;
 import cc.lq.blog.system.mapper.CategoryMapper;
@@ -50,18 +51,23 @@ public class ArticleServiceImpl extends ServiceImpl<ArticleMapper, ArticleDO> im
     public ArticleVO getArticleVOByArticleId(Long articleId) {
         ArticleDO articleDO = this.articleMapper.selectById(articleId);
         if (articleDO == null) {
-            return null;
+            throw new ResourceNotFoundException("未找到id为" + articleId + "的文章");
         }
-        CategoryDO category = this.categoryMapper.selectById(articleDO.getCategoryId());
+        CategoryDO categoryDO = this.categoryMapper.selectById(articleDO.getCategoryId());
+        CategoryVO categoryVO = new CategoryVO();
+        BeanUtils.copyProperties(categoryDO, categoryVO);
         List<ArticleTagDO> articleTagDOList = this.articleTagMapper.selectList(
                 new QueryWrapper<ArticleTagDO>().eq("article_id", articleId));
-        List<TagDO> tagList = new ArrayList<>();
-        articleTagDOList.forEach(articleTagDO ->
-                tagList.add(this.tagMapper.selectById(articleTagDO.getTagId()))
+        List<TagVO> tagList = new ArrayList<>();
+        articleTagDOList.forEach(articleTagDO -> {
+                    TagDO tagDO = this.tagMapper.selectById(articleTagDO.getTagId());
+                    TagVO tagVO = new TagVO(tagDO.getId(), tagDO.getTagName());
+                    tagList.add(tagVO);
+                }
         );
 
         ArticleVO articleVO = new ArticleVO();
-        articleVO.setCategory(category);
+        articleVO.setCategory(categoryVO);
         articleVO.setTags(tagList);
         BeanUtils.copyProperties(articleDO, articleVO);
 
@@ -70,6 +76,10 @@ public class ArticleServiceImpl extends ServiceImpl<ArticleMapper, ArticleDO> im
 
     @Override
     public Boolean removeArticleByArticleId(Long articleId) {
+        ArticleDO articleDO = this.articleMapper.selectById(articleId);
+        if(articleDO == null) {
+            throw new ResourceNotFoundException("未找到id为" + articleId + "的文章");
+        }
         int result = this.articleMapper.deleteById(articleId);
         QueryWrapper<ArticleTagDO> wrapper = new QueryWrapper<>();
         wrapper.eq("article_id", articleId);
@@ -95,46 +105,45 @@ public class ArticleServiceImpl extends ServiceImpl<ArticleMapper, ArticleDO> im
         }
         articleVO.setId(articleDO.getId());
 
-        return articleVO;
+        return getArticleVOByArticleId(articleVO.getId());
     }
 
     @Override
     public ArticleVO updateArticleVO(ArticleVO articleVO) {
         if (articleVO.getId() == null) {
-            throw new IllegalArgumentException();
+            throw new IllegalArgumentException("id不能为空");
         }
         if (articleVO.getCategory() == null || articleVO.getCategory().getId() == null) {
-            throw new IllegalArgumentException();
+            throw new IllegalArgumentException("类别或类别id不能为空");
         }
         Long categoryId = articleVO.getCategory().getId();
         if (this.categoryMapper.selectById(categoryId) == null) {
-            throw new IllegalArgumentException();
+            throw new ResourceNotFoundException("找不到id为" + categoryId + "的类别");
         }
 
-        ArticleService articleServiceImpl = this.applicationContext.getBean("articleServiceImpl", ArticleService.class);
-        ArticleVO originalArticleVO = articleServiceImpl.getArticleVOByArticleId(articleVO.getId());
+        ArticleVO originalArticleVO = getArticleVOByArticleId(articleVO.getId());
         if (originalArticleVO == null) {
-            return null;
+            throw new ResourceNotFoundException("找不到id为" + categoryId + "的文章");
         }
 
         ArticleDO articleDO = new ArticleDO();
         BeanUtils.copyProperties(articleVO, articleDO);
         articleDO.setCategoryId(articleVO.getCategory().getId());
         this.articleMapper.updateById(articleDO);
-        List<TagDO> tags = articleVO.getTags();
+        List<TagVO> tags = articleVO.getTags();
 
         QueryWrapper<ArticleTagDO> wrapper = new QueryWrapper<>();
         wrapper.eq("article_id", articleVO.getId());
         this.articleTagMapper.delete(wrapper);
         if (tags != null && tags.size() > 0) {
-            tags.forEach(tagDO -> {
+            tags.forEach(tagVO -> {
                 ArticleTagDO articleTagDO = new ArticleTagDO();
                 articleTagDO.setArticleId(articleVO.getId());
-                articleTagDO.setTagId(tagDO.getId());
+                articleTagDO.setTagId(tagVO.getId());
                 this.articleTagMapper.insert(articleTagDO);
             });
         }
-        return articleVO;
+        return getArticleVOByArticleId(articleVO.getId());
     }
 
     @Override
@@ -153,7 +162,11 @@ public class ArticleServiceImpl extends ServiceImpl<ArticleMapper, ArticleDO> im
         QueryWrapper<ArticleDO> wrapper = new QueryWrapper<>();
         wrapper.eq("category_id", categoryId)
                 .orderByDesc("update_time").orderByDesc("create_time");
-        return this.articleMapper.selectPage(page, wrapper);
+        Page<ArticleDO> articleDOPage = this.articleMapper.selectPage(page, wrapper);
+        if(articleDOPage.getRecords().size() == 0) {
+            throw new ResourceNotFoundException("未找到指定条件的文章");
+        }
+        return articleDOPage;
     }
 
     @Override
